requirements.txt
python==3.11.9
langchain==0.1.19
langchain-openai==0.1.3
python-dotenv==1.0.1
pandas==2.1.4
SpeechRecognition==3.10.4
Django==5.1
PyAudio==0.2.14
chromadb==0.4.24
docx2txt==0.8